requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.2.0
attrs==23.2.0
bitarray==2.6.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crc16==0.1.1
crc32c==2.3.post0
fastapi==0.109.2
frozenlist==1.4.1
graphql-core==3.2.3
graphql_query==1.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
pycparser==2.21
pydantic==1.10.14
PyNaCl==1.5.0
python-dotenv==1.0.1
pytonlib==0.0.53
PyYAML==6.0.1
requests==2.31.0
setuptools==69.1.0
sniffio==1.3.0
starlette==0.36.3
ton==0.26
tonsdk==1.0.13
TonTools==2.1.2
tvm-valuetypes==0.0.10
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
